GH-1348 Disallow overriding AutoConfiguration beans in JwtAuth, AxelixConfigurationsProperties, and AxelixEnvironmentEndpoint in spring-boot-3
