+++ sbs/axelix-spring-boot-3-starter/src/main/java/com/axelixlabs/axelix/sbs/spring/autoconfiguration/AxelixConfigurationsPropertiesEndpointAutoConfiguration.java
@@ -19,14 +19,10 @@ package com.axelixlabs.axelix.sbs.spring.autoconfiguration;
 
 import org.springframework.boot.actuate.autoconfigure.endpoint.condition.ConditionalOnAvailableEndpoint;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 
-import com.axelixlabs.axelix.common.auth.core.SecurityContextExecutor;
 import com.axelixlabs.axelix.sbs.spring.core.auth.RequiredAuthorityCheckService;
-import com.axelixlabs.axelix.sbs.spring.core.auth.ThreadLocalSecurityContextExecutor;
-import com.axelixlabs.axelix.sbs.spring.core.config.EndpointsConfigurationProperties;
 import com.axelixlabs.axelix.sbs.spring.core.configprops.AxelixConfigurationPropertiesEndpoint;
 import com.axelixlabs.axelix.sbs.spring.core.configprops.ConfigurationPropertiesConverter;
 import com.axelixlabs.axelix.sbs.spring.core.configprops.ConfigurationPropertiesFlattener;
@@ -35,8 +31,6 @@ import com.axelixlabs.axelix.sbs.spring.core.configprops.DefaultConfigurationPro
 import com.axelixlabs.axelix.sbs.spring.core.configprops.DefaultConfigurationPropertiesFlattener;
 import com.axelixlabs.axelix.sbs.spring.core.configprops.DefaultConfigurationPropertiesService;
 import com.axelixlabs.axelix.sbs.spring.core.configprops.SmartSanitizingFunction;
-import com.axelixlabs.axelix.sbs.spring.core.env.DefaultPropertyNameNormalizer;
-import com.axelixlabs.axelix.sbs.spring.core.env.PropertyNameNormalizer;
 
 /**
  * Auto-configuration for the {@link AxelixConfigurationPropertiesEndpoint}.
@@ -44,53 +38,22 @@ import com.axelixlabs.axelix.sbs.spring.core.env.PropertyNameNormalizer;
  * @since 13.11.2025
  * @author Sergey Cherkasov
  */
-@AutoConfiguration(after = EndpointsConfigurationPropertiesAutoConfiguration.class)
+@AutoConfiguration(after = EndpointPropertiesSupportAutoConfiguration.class)
 @ConditionalOnAvailableEndpoint(endpoint = AxelixConfigurationPropertiesEndpoint.class)
 public class AxelixConfigurationsPropertiesEndpointAutoConfiguration {
 
     @Bean
-    @ConditionalOnMissingBean
     public ConfigurationPropertiesFlattener configurationPropertiesFlattener() {
         return new DefaultConfigurationPropertiesFlattener();
     }
 
     @Bean
-    @ConditionalOnMissingBean
     public ConfigurationPropertiesConverter configurationPropertiesConverter(
             ConfigurationPropertiesFlattener configurationPropertiesFlattener) {
         return new DefaultConfigurationPropertiesConverter(configurationPropertiesFlattener);
     }
 
     @Bean
-    @ConditionalOnMissingBean
-    public PropertyNameNormalizer propertyNameNormalizer() {
-        return new DefaultPropertyNameNormalizer();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public SmartSanitizingFunction smartSanitizingFunction(
-            EndpointsConfigurationProperties endpointsConfigurationProperties,
-            PropertyNameNormalizer propertyNameNormalizer) {
-        return new SmartSanitizingFunction(
-                endpointsConfigurationProperties.getSanitizedProperties(), propertyNameNormalizer);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public SecurityContextExecutor securityContextExecutor() {
-        return new ThreadLocalSecurityContextExecutor();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public RequiredAuthorityCheckService requiredAuthorityCheckService(
-            SecurityContextExecutor securityContextExecutor) {
-        return new RequiredAuthorityCheckService(securityContextExecutor);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
     public ConfigurationPropertiesService configurationPropertiesService(
             SmartSanitizingFunction smartSanitizingFunction,
             ApplicationContext applicationContext,
@@ -104,7 +67,6 @@ public class AxelixConfigurationsPropertiesEndpointAutoConfiguration {
     }
 
     @Bean
-    @ConditionalOnMissingBean
     public AxelixConfigurationPropertiesEndpoint axelixConfigurationPropertiesEndpoint(
             ConfigurationPropertiesService configurationPropertiesService) {
         return new AxelixConfigurationPropertiesEndpoint(configurationPropertiesService);
